fix(runner): retire the LLM tool-call deadline
An LLM tool call backed by a pattern created a result cell, started the
pattern against it, and then waited for that cell to become defined.
The wait was value-shaped, and a value cannot answer the question being
asked. "The result cell became defined" conflates two things — has the
run finished, and did it produce anything — so a run that finishes
having written nothing looks exactly like one still working, and there
is no moment at which the wait can conclude that nothing is coming.

A 120-second deadline filled that gap. It bounded how long a healthy
local computation was allowed to take, which is the shape the
repository's engineering principles rule out: anything that would have
completed a second later cannot complete at all, and the model was
handed `{ type: "error-text", value: "Tool call timed out" }` in place
of a result the tool was about to produce. The failure was invisible in
tests, because the conversation continues either way.

The wait now races three events and no clock: the result landing, the
turn's abort signal, and `Runtime.settledFor` reporting that the run
this tool call started has gone quiet. A tool that produced nothing
ends through the third with its cell still undefined, which is a
determinate answer rather than a guess. Nesting composes without a
lineage graph: a nested tool call's own wait sits inside work
registered against the child, so waiting on the child transitively
covers the grandchild.

`REQUEST_TIMEOUT` stays. It bounds how long to keep believing a
different replica is still working on a dialog, which is failure
detection across a network and cannot be made exact.

## The registry the barrier reads had a hole in it

The runtime keeps a registry of in-flight asynchronous builtin work,
and it is the only thing these barriers can see. `generateObject` never
reached it. Both of its paths built a promise spanning the model call
and the writeback and then attached only a `.catch()` to it. An
instrumented run that logged every registration while driving the
dynamic-schema subagent tests recorded four registrations across the
whole file, all of them commit promises from the scheduler, and nothing
at all from `generateObject`. Since `generateObject` is the path an LLM
tool call runs, a barrier built on that registry would have reported
the child agent finished before it had made its request — reporting
success with the data missing, which is worse than the deadline it
replaces, because a deadline firing early at least reports a failure.

The direct path needed care beyond adding a call: its writeback hangs
off a `.then()` on the request promise, so the chain rather than the
request is what spans the operation. Every other async builtin already
registered. Two enqueue post-commit effects and deliberately do not:
the sqlite query awaits its RPC and writeback inside the flush, so the
transaction's own commit promise spans them, and `streamData` is an
open-ended subscription whose read loop has no completion for a barrier
to wait for.

Registration alone is not enough for a barrier scoped to one pattern
run, which is what a tool call needs — it cannot wait for the whole
runtime, because it is itself running inside registered work and would
wait on the promise it is inside. So each registration now carries the
run it belongs to. The link was already threaded everywhere and needed
no new plumbing: a pattern's nodes are all instantiated against the
same result cell, and that cell reaches every raw builtin as its
`parentCell`, so a builtin naming `parentCell` attributes its work to
its own run.

`settledFor` waits for unowned work as well, and that is not a loose
end. The scheduler registers the commit promise of any commit carrying
post-commit effects, and that promise is the handoff: it spans the
outbox flush inside which a builtin registers its own work. Without it
a run-scoped barrier slips through the gap between the commit landing
and the flush starting, and finds an empty set. It also carries no
round cap, where the whole-runtime barrier does: on a barrier a tool
call waits on, a cap is a bound on how much work a run may do before
the barrier reports it finished anyway, which is the same shape as the
deadline being removed. Each round awaits real promises, so a run that
keeps working keeps the barrier open rather than spinning. An unbounded
loop does need a way to stop, or a race the result wins leaves it
re-checking with nobody waiting, so it takes a signal the tool call
aborts as it leaves the race.

## Guarding the registry, and testing the barrier honestly

`async-work-registration.test.ts` guards the registry in two halves.
Each async builtin runs there with its response held open; both
barriers must stay open while it is held, and a barrier scoped to an
unrelated cell must return, so the pair pins that the work was
registered and that it was attributed to its own run rather than to
nobody. A source scan in the same file fails when a builtin grows a
post-commit side effect without registering async work, so a builtin
added later has to either register or record why it does not have to.
Both halves were checked against the defect: removing the
`generateObject` registrations turns the first red, and dropping the
owner turns all four cases red.

The tool-calling tests could not tell an early return from a correct
one, because every mock in them answers within a microtask and a child
agent finishes before its tool call looks at anything. A new case holds
a delegate's child answer open across the whole wait and checks that
the parent's next request carries the child's data rather than an empty
result. Removing the barrier from the race turns it red.

## Retiring the real-clock exemptions

Three tool-calling cases were split into their own files and kept on
the real clock, because their delegate could not form the child's
request until the model-supplied schema had settled through the graph,
and that round trip carried the delegate's completion across a
macrotask boundary. The fake clock's auto-advance pump read that
boundary as an idle event loop and jumped logical time to the earliest
pending production timer, which was the deadline. The exemption could
not be traded for a smarter pump: the pump cannot tell that deadline
apart from a backoff window, and the backpressure tests need backoff
windows to fire during exactly this kind of reactive churn.

With the deadline gone the cases return to their parent files, both
split files are deleted, and both entries leave the preload's
real-clock list, leaving only the resume test that is there for an
unrelated reason. This proves itself rather than being taken on trust:
each case asserts on the delegate's own tool result, and the difference
is otherwise invisible because the conversation reaches its closing
assertion either way.

## What this gives up, stated rather than hidden

A userland pattern that spins forever with nobody watching now leaves
its turn pending instead of being abandoned after two minutes. That
trades misreporting a working tool as failed for a turn that stays
pending until the work finishes or someone cancels, and cancellation
reaching tool calls is what makes the second a real option.

One sharp edge is unchanged and now recorded: the result-landing racer
fires on the result cell becoming defined at all, so a delegate pattern
returning the whole builtin object rather than its `.result` resolves
its tool call on a `{ pending: true }` placeholder no matter what the
barrier does.
