+++ packages/runner/src/builtins/navigate-to.ts
@@ -82,8 +82,9 @@ export function navigateTo(
         }
       });
       // Navigation is an external effect: release it only after a successful
-      // commit. The outbox promise is tracked explicitly so runtime.settled()
-      // cannot race async shell navigation.
+      // commit. The outbox promise is tracked explicitly, owned by this run, so
+      // neither runtime.settled() nor runtime.settledFor(parentCell) can race
+      // async shell navigation.
       const targetLink = resolvedTarget.getAsNormalizedFullLink();
       tx.enqueuePostCommitEffect({
         // The outbox deduplicates by id within a transaction. Encode the full
@@ -103,7 +104,7 @@ export function navigateTo(
           const work = Promise.resolve().then(() =>
             navigateCallback(resolvedTarget)
           );
-          runtime.trackAsyncWork(work);
+          runtime.trackAsyncWork(work, parentCell);
           try {
             await work;
           } catch (error) {
