+++ packages/runner/test/clock-preload.ts
@@ -22,19 +22,5 @@ installFakeClock({
     // Under the fake clock the delay is a frozen test-file timer that pull/idle
     // wait on, so the resume deadlocks; the real clock delivers them as intended.
     "list-resume-preserve",
-    // A generateObject delegate tool runs a child agent against a result schema
-    // the model supplies in the tool input, so the child reaches its own request
-    // only after that input has settled through the graph. The tool-calling path
-    // guards its wait with a deadline, and the pump reads the resulting macrotask
-    // boundary as an idle loop and jumps to that deadline, aborting the delegate
-    // ("Tool call timed out") mid-flight. The pump cannot tell that deadline from
-    // a backoff window, which other tests need it to fire during the same churn.
-    // Retiring that deadline is what lets this file and the next one convert.
-    // Each case asserts on the delegate's own tool result, so dropping an entry
-    // here turns them red rather than quietly green.
-    "generate-object-tools-dynamic-subagent",
-    // The same delegate-tool shape, reached through llmDialog rather than
-    // generateObject.
-    "llm-dialog-dynamic-subagent",
   ],
 });
