+++ packages/runner/test/llm-dialog.test.ts
@@ -2080,6 +2080,182 @@ describe("llmDialog", () => {
       await ceilingStorageManager.close();
     }
   });
+  it("should expose a userland subagent in llmDialog tool catalogs", async () => {
+    const childResultSchema = {
+      type: "object",
+      properties: {
+        approved: { type: "boolean" },
+        summary: { type: "string" },
+      },
+      required: ["approved", "summary"],
+      additionalProperties: false,
+    } as const satisfies JSONSchema;
+
+    loadConversationFixture({
+      description: "llmDialog subAgent tool should be available to the parent",
+      responses: [
+        {
+          type: "sendRequest",
+          expectRequest: {
+            hasTools: ["delegate"],
+            messageCount: 1,
+          },
+          response: {
+            role: "assistant",
+            content: [{
+              type: "tool-call",
+              toolCallId: "call_delegate",
+              toolName: "delegate",
+              input: {
+                prompt: "analyze the hidden text",
+                resultSchema: childResultSchema,
+              },
+            }],
+            id: "dlg-subagent-r1",
+          },
+        },
+        {
+          type: "sendRequest",
+          expectRequest: {
+            hasTools: ["helperTool", "presentResult"],
+            messageCount: 1,
+          },
+          response: {
+            role: "assistant",
+            content: [{
+              type: "tool-call",
+              toolCallId: "call_child_present",
+              toolName: "presentResult",
+              input: {
+                approved: false,
+                summary: "Not approved.",
+              },
+            }],
+            id: "dlg-subagent-r2",
+          },
+        },
+        {
+          type: "sendRequest",
+          expectRequest: {
+            hasTools: ["delegate"],
+            messageCount: 3,
+          },
+          response: {
+            role: "assistant",
+            content: "Delegate completed.",
+            id: "dlg-subagent-r3",
+          },
+        },
+      ],
+    });
+
+    const helperTool = pattern(
+      () => ({ ok: true }),
+      {
+        type: "object",
+        additionalProperties: false,
+      },
+      {
+        type: "object",
+        properties: {
+          ok: { type: "boolean" },
+        },
+        required: ["ok"],
+        additionalProperties: false,
+      } as const satisfies JSONSchema,
+    );
+
+    const resultSchema = {
+      type: "object",
+      properties: {
+        addMessage: { ...LLMMessageSchema, asCell: ["stream"] },
+        pending: { type: "boolean" },
+        messages: {
+          type: "array",
+          items: { type: "object", additionalProperties: true },
+        },
+      },
+      required: ["addMessage"],
+    } as const satisfies JSONSchema;
+
+    const subAgentPattern = pattern<any, any>(
+      ({ prompt, resultSchema }) => {
+        return generateObject({
+          prompt,
+          schema: resultSchema,
+          tools: {
+            helperTool: patternTool(
+              helperTool,
+            ) as unknown as BuiltInLLMTool,
+          },
+        } as any).result;
+      },
+      {
+        type: "object",
+        properties: {
+          prompt: { type: "string" },
+          resultSchema: {
+            type: "object",
+            additionalProperties: true,
+          },
+        },
+        required: ["prompt", "resultSchema"],
+        additionalProperties: false,
+      },
+      true,
+    );
+
+    const testPattern = pattern(
+      () => {
+        const messages = Cell.of<BuiltInLLMMessage[]>([]);
+        const dialog = llmDialog({
+          messages,
+          tools: {
+            delegate: {
+              description: "Run a child agent and return schema-limited JSON.",
+              ...(patternTool(subAgentPattern) as unknown as BuiltInLLMTool),
+            },
+          },
+        });
+        return {
+          addMessage: dialog.addMessage,
+          pending: dialog.pending,
+          messages,
+        };
+      },
+      false,
+      resultSchema,
+    );
+
+    const resultCell = runtime.getCell(
+      space,
+      "llmDialog-subagent-tool-test",
+      resultSchema,
+      tx,
+    );
+
+    const result = runtime.run(tx, testPattern, {}, resultCell);
+    tx.commit();
+
+    const addMessage = await result.key("addMessage").pull();
+    addMessage.send({ role: "user", content: "Start the workflow." });
+
+    await waitForMessages(result, 4);
+    const messages = (await result.key("messages").pull())!;
+    // The conversation reaches its third request whether the delegate returned
+    // data or an error, so the closing assertion below cannot see an abandoned
+    // delegate. Assert on the tool result itself: a completed delegate reports
+    // `{ type: "json", ... }`, an abandoned one reports
+    // `{ type: "error-text", ... }`.
+    const toolMessage = messages.find((message: any) =>
+      message.role === "tool"
+    );
+    const delegateOutput = (toolMessage?.content as any[])?.find((part) =>
+      part?.type === "tool-result" && part.toolName === "delegate"
+    )?.output;
+    expect(delegateOutput).toMatchObject({ type: "json" });
+    expect(messages.at(-1)?.content).toBe("Delegate completed.");
+  });
 });
 
 function waitForMessages(result: any, expectedCount: number) {
